LayoutTests/imported/w3c/web-platform-tests/css/css-sizing/responsive-iframe/responsive-iframe-icb.tentative-expected.txt



FAIL IFRAME content should use the initial ICB to measure the natural size assert_equals: expected 302 but got 152
